Add optional sanity checks.
Limited slider spam to not generate a million notes.

Add a general per-beatmap error catch net, so that individual beatmap
errors cannot stop the general process.

Fixed simfiles leaking from beatmapset to beatmapset if they were not
consumed. A notable example of non-consumption is when an error arises.

Do not skip inherited timing points now that their info is actually
used.

Handle slider length as the length of the entire slider rather than a
single slide over it.

Use the last slider control point as the reference last position
(instead of the start point).