Simplify `in_place` to a boolean flag
Now `SimfileWrite` creates/checks the symlink lazily on the first
simfile written.

Errors at write-time if an attempt to write simfiles from multiple
sources is made.

Fix that dependencies in subdirectories would fail to copy because the
directory would not be present in the output location.

Do not abort the entire beatmapset if a dependency fails to copy.

Added a last-resort `CopyMethod` that only asserts that source and
destination are identical.
This fixes issues with making hardlinks over hardlinks that already
exist.

Prefer symlinks on Unix systems by default.